Make MessageDeleted#previous_message optional
Slack's API allows for an optional previous_message, because
of the following fun scenario:

When the OP in a thread is deleted, the thread remains alive.
The thread still needs to be attached to something; this "something"
is still a message, but a soft-deleted message they call a "tombstone."

When all the threaded messages are deleted, a callback automatically
deletes the last message; Slack will still send out a webhook event for
this deletion, but because this "tombstone" deletion no longer has a
previous message, it can be nil.

¯\_(ツ)_/¯